Refactoring - combined preprocessing module
Moved various functions into a preprocessing module. Closes #29
